add article class file
